+++ MovieClubManager.Service/Movies/Contracts/Dto/GetMovieDto.cs
@@ -11,10 +11,10 @@ namespace MovieClubManager.Service.Movies.Contracts.Dto
         public int Id { get; set; }
         public string Name { get; set; }
         public string PublishYear { get; set; }
-        public int DailyPriceRent { get; set; }
+        public decimal DailyPriceRent { get; set; }
         public int AgeLimit { get; set; }
-        public int DelayPenalty { get; set; }
-        public double Duration { get; set; }
+        public decimal DelayPenalty { get; set; }
+        public decimal Duration { get; set; }
         public string Director { get; set; }
         public string? Description { get; set; }
         public int GenreId { get; set; }
